requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pytest==3.0.0
Werkzeug==0.11.4
python-dateutil==2.5.3